backend/requirements.txt
# Geospatial
shapely==2.0.6

# API client
requests==2.32.3

# Utilities
python-dotenv==1.0.1

# Testing
pytest==8.2.0
